+++ src/lib.rs
@@ -13,12 +13,11 @@ pub(crate) mod slice;
 
 pub mod prelude {
     pub use crate::animation::{
-        Animation, AnimationDirection, AnimationEvents, AnimationRepeat, AnimationState,
-        AseSpriteAnimation, AseUiAnimation,  ManualTick, NextFrameEvent,
-        PlayDirection, AseAnimation, MaterialAnimationPlugin
+        render_animation, Animation, AnimationDirection, AnimationEvents, AnimationRepeat,
+        AnimationState, AseAnimation, ManualTick, NextFrameEvent, PlayDirection, RenderAnimation,
     };
-    pub use crate::loader::{Aseprite, AsepriteLoaderSettings};
-    pub use crate::slice::{AseSpriteSlice, AseUiSlice};
+    pub use crate::loader::{Aseprite, AsepriteLoaderSettings, SliceMeta};
+    pub use crate::slice::{render_slice, AseSlice, RenderSlice};
     pub use crate::AsepriteUltraPlugin;
 }
 
@@ -76,7 +75,3 @@ impl Plugin for AsepriteUltraPlugin {
         app.add_plugins(processor::AsepriteProcessorPlugin);
     }
 }
-
-/// component to signal a aseprite render is fully loaded.
-#[derive(Component, Default)]
-pub(crate) struct FullyLoaded;